Add files via upload
This is another assignment from software development 1. Worked on creating a game that ran in a loop which gave the "player" in this case the possibility to be the 'Guesser' or the 'Setter' to which the Guesser has 7 opportunities to guess a random number generated from 1-100 inclusive. As by the Setter, the player inputs a number, again from 1-100 inclusive, and the computer has 7 opportunities to guess the number. The only way to quit the game is by inputting "Q" when asked what the player wants to be. The code can be messed with to change the range.