Add files via upload
This is the classic tic tac toe game. Simple assignments, quite fun to code. Its a 3x3 board as classic but the cells go from 1-9 in an awkward style. I can't think of a way to make it more user-friendly as of today but It works perfectly fine and its as fun as a tic tac toe game can be.