Add files via upload
This is my first ever assignment and a proper program that calculates the temperature of a coffee mug sitting in an office according to the starting time and temperature. Honestly, this one is more for me to see in the future and see how much better of a programmer I have become.